Updated the design from CharacterForm, Landing and SelectCharacter pages
