+++ web/src/routing.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { BrowserRouter, Route } from "react-router-dom"
 
 import CharacterDetails from "./pages/CharacterDetails"
-import CreateCharacter from "./pages/CreateCharacter"
+import CharacterForm from "./pages/CharacterForm"
 import Landing from "./pages/Landing"
 import SelectCharacter from "./pages/SelectCharacter"
 
@@ -11,7 +11,7 @@ export default function Router() {
     return (
         <BrowserRouter>
             <Route path="/" exact component={Landing} />
-            <Route path="/create" exact component={CreateCharacter} />
+            <Route path="/create" exact component={CharacterForm} />
             <Route path="/select" exact component={SelectCharacter} />
             <Route path="/character/:id" exact component={CharacterDetails} />
         </BrowserRouter>
