+++ web/src/styles/components/characterThumbnail.css
@@ -1,5 +1,5 @@
 .character-thumb {
-    background-color: gainsboro;
+    background-color: black;
     display: flex;
     flex-direction: row;
     justify-content: center;
@@ -7,11 +7,26 @@
 }
 
 .character-thumb:hover {
-    background-color: rgb(201, 198, 198);
+    background-color: rgb(0, 78, 102);
     cursor: pointer;
 }
 
+.character-thumb-wrapper img {
+    border-radius: 100%;
+    width: 120px;
+    height: 120px;
+    padding-left: 10px;
+    padding-top: 10px;
+}
+
+.add-character-thumb {
+    background-color: black;
+    color: white;
+    border: none;
+}
+
 .delete-icon-wrapper {
+    background-color: black;
     width: min-content;
     height: min-content;
     border: none;
