+++ web/src/styles/global.css
@@ -1,11 +1,25 @@
-body{
-    background-color: #392F1F;
+@import url('https://fonts.googleapis.com/css2?family=Open+Sans&family=Roboto+Condensed:wght@300&family=Roboto+Slab&display=swap');
+:root {
+    --main-bg-color: #03045E;
+    --bg-color: #1c1d64;
+    --button-color: #0077B6;
+    --button-color-hover: #307aa1;
+    --main-text-color: #FFFFFF;
+    --input-background: #05071a;
+    /* --main-color: #00B4D8;
+    --main-color: #00B4D8; */
+}
+
+body {
+    background-color: var(--main-bg-color);
+    color: var(--main-text-color);
     display: flex;
     align-items: center;
     justify-content: center;
+    font-family: 'Roboto Condensed', sans-serif;
 }
 
-img.character-img{
+img.character-img {
     max-width: 150px;
     max-height: 150px;
 }
