+++ web/src/styles/pages/characterForm.css
@@ -1,0 +1,67 @@
+.create-character-form .character-image .character-img-input {
+    width: 0;
+    height: 0;
+    visibility: hidden;
+}
+
+.create-character-form {
+    display: flex;
+    flex-direction: row;
+    font-size: larger;
+}
+
+
+.create-character-form .character-image-preview img {
+    border-radius: 100%;
+    width: 200px;
+    height: 200px;
+}
+
+.create-character-form fieldset {
+    border: hidden;
+    background-color: var(--bg-color);
+}
+
+@media all and (max-width: 500px) {
+    .create-character-form {
+        flex-direction: column;
+    }
+}
+
+
+.edit-icon :hover {
+    cursor: pointer;
+}
+
+.character-image-preview {
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+    width: 50%;
+}
+
+
+.character-name {
+    flex-direction: column;
+    display: flex;
+}
+
+.character-description {
+    margin-top: 20px;
+    display: flex;
+    flex-direction: column;
+}
+
+.character-submit-button {
+    margin-top: 20px;
+    margin-right: auto;
+    margin-left: auto;
+    padding: 20px;
+    border: none;
+    color: #ffffff;
+    background-color: var(--button-color);
+}
+
+.character-submit-button:hover {
+    background-color: var(--button-color-hover);
+}
