--- web/src/styles/pages/createCharacter.css
@@ -1,43 +1,0 @@
-.create-character-form .character-image .character-img-input {
-    width: 0;
-    height: 0;
-    visibility: hidden;
-}
-
-.edit-icon {
-    position: relative;
-    width: 27px;
-    height: 27px;
-}
-
-.edit-icon :hover {
-    cursor: pointer;
-}
-
-.character-image-preview {
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-    width: 50%;
-}
-
-.create-character-form fieldset {
-    border: hidden;
-    background-color: rgb(122, 88, 37, 0.5);
-}
-
-.character-name {
-    flex-direction: column;
-    display: flex;
-}
-
-.character-description {
-    margin-top: 20px;
-    display: flex;
-    flex-direction: column;
-}
-
-.character-submit-button {
-    margin-right: auto;
-    margin-left: auto;
-}
