+++ web/src/styles/pages/landing.css
@@ -1,21 +1,49 @@
-.title {
+.page-wrapper {
+    color: #FFFFFF;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    padding: 10px;
+}
+
+
+
+@media all and (max-width: 300px) {
+    .page-wrapper {
+        flex-direction: column;
+    }
+    .about{
+        display: flex;
+        flex-direction: row;
+    }
+}
+
+.about .title {
     font-size: 50;
 }
 
 .link-wrapper {
-    padding: 20px;
-    background-color: #C4C4C4;
+    background-color: var(--button-color);
+    border: none;
     display: flex;
-    justify-content: center;
     align-items: center;
+    justify-content: center;
+    min-width: 200px;
+    height: 70px;
 }
 
-.link-wrapper + .link-wrapper{
-    margin-top: 20px;
+.link-wrapper:hover {
+    cursor: pointer;
+    background-color: var(--button-color-hover);
 }
 
-.link {
-    font-size: larger;
+.link-wrapper .link {
     text-decoration: none;
+}
+
+.link-wrapper .link .link-text {
+    color: #FFFFFF;
+    font-size: larger;
+    width: 100%;
     height: 100%;
 }
